+++ src/store/layout/layoutReducer.js
@@ -5,7 +5,7 @@ const initialState = {
 export default function(state = initialState, action) {
   const { type } = action;
 
-  if (type === 'LAYOUT_MENU_TOGGLE') {
+  if (type === 'LAYOUT-MENU-TOGGLE') {
     return {
       ...state,
       isMenuVisible: !state.isMenuVisible,